Polling implemented
Used a kernel timer to poll every RX_POLL_INTERVAL jiffies for packets. Used NAPI to slurp up all t

Used a kernel timer to poll every RX_POLL_INTERVAL jiffies for packets,
and used NAPI to slurp up all the received packets when the polling
mechanism sees there are one or more to receive. Used napi_struct,
registered with netif_napi_add. Found that I needed to napi_add the
napi_struct before registering the net_device nf10_netdev... otherwise
the kernel dereferences a NULL pointer in netif_napi_add. Use
napi_schedule and napi_complete to schedule and stop polling,
respectively. add_timer takes care of re-adding the polling timer when
the timer fires.

Few things to look into still:
 - ARP packets seem to be below the reported size of 60 (42), so need to
check out this lower bound on packet sizes to see if its really
necessary.
 - Want to make sure that netif_receive_skb is really the right function
to be calling during the NAPI packet slurping.
 - Want to make sure to disable the rx_poll_timer in the right way when
we stop() the interface.
 - Noticed skull code that sets the ip_summed field of the skb on packet
receiving to NO_CHECKSUMMING or something like that.... probably want to
look into whether or not we want to set that field of the skb. Currently
it is just not set.
 - Experienced a strange crash after long up-time (10-15 minutes) of the
driver. Not really sure what's going on there... will need to try and
reproduce.

Todo:
 - Update all the statistics counters correctly.
 - For when running the actual NIC hardware, need to pull out the MAC
addresses of the phys
 - Need to replicate 4x the netdevs
   - Need to understand whatever trick we decided on for multiplexing
the DMA path.
 - Write more unit tests!