Got rid of printfs in the register access library.
Shouldn't have printf's in the library... so I removed them, and then I also replaced the error reporting so that it uses the netlink error messages NLE_* from include/netlink/errno.h... which seems to be included from the other header files that I used so I didn't need to include that specifically. As far as I can tell, all the error message reporting is now exclusively from that namespace.