Changed read register interface to report errors.
Now the return value of nf10_reg_rd is the error code, and there's another argument which you use to send it a pointer to the val variable you want set.

Also I fixed some error reporting problem. nl_recvmsgs_default was returning the value of the register, but could have been confounded with an error code. So instead I now send the callback function a pointer to the val variable that the caller wants set, so that it can set it, and return a pure error code. Now nl_recvmsgs_default sends us back only an error code, which I observe and send back if necessary.