BUG! Receiving bad opcode crashes the kernel.
In the nap_struct_poll() function, it tries to interpret the dst_iface from the received opcode via a call to something like rx_get_dst_iface(), but the return value of that function is not checked. The function does return -1 if it can't determine the dst_iface, and becaues of the unchecked return value the kernel ends up using -1 as an index into the netdev array, thus crashing the kernel.

Also starting looking through for other unchecked iface indexs...