Added error reporting between kernel and library.
driver now reports back an errno attribute to the library, so that the library now can know whether or not the operation really succeeded once it handed it off to the driver. Before this modification , register writes were assumed to succeed if everything went smooth in the library. Register reads were assumed to have worked as long as we got back a read value... so there was just no solid way of seeing errors come from the driver. Now there is