Solved no-hardware kernel crash issue.
Solved the issue of the driver crashing the computer when there's no hardware installed. This was because pci_register_driver would actually return 0 even when no hardware was found... but I assumed that this would not happen. Therefore, the rx_poll_timer was started, and the poll call back queries for packets using variables that are only intialized if probe() had been called... so I moved the timer starting into the end of the probe() function, and moved pci_driver registration to the bottom of _init so that all th netdevs and other data structures can be properly allocated and initialized before probe starts the timer.